Add password hashing and checking logic; integrate with signup, login, and Account
