+++ main.py
@@ -1,4 +1,5 @@
 from models.UserGate import UserGate
+from models.Account import Account
 import services.user_services
 from repositories import read_write_userID
 import repositories.read_write_User_data
@@ -19,8 +20,10 @@ def first_menu():
 
 
 acc1 = UserGate()
-acc1 = acc1.login("rabbit5", "carrot")
+acc1 = acc1.login("coolguy123", "pass123")
+#print(isinstance(acc1, Account))
+#print(acc1.deposit(1000))
 #print(acc1.buy(1001))
-print(acc1.view_order_history)
-print(acc1.account_balance)
+#print(acc1.view_order_history)
+#print(acc1.account_balance)
 #print(acc1.add_to_cart(1006))
